# HDDS-8773. [S3G] Improve list performance in FSO bucket